requirements.txt
haystack-ai
haystack-experimental
ollama-haystack
deepeval-haystack
hayhooks
markdown-it-py 
mdit_plain
PyYAML
pypdf
sentence-transformers
huggingface_hub
python-dotenv
python-telegram-bot
flask
flask_sqlalchemy